requirements.txt
faiss-cpu
sentence-transformers
pymongo
fastapi
uvicorn
bson
numpy
python-multipart
mangum
python-dotenv
transformers
Pillow
